Began work on an edit grades page.
Implemented button that goes to a page, with all the necessary model
information. Just make it look nice and have a text box for the score,
and a button that calls a submit or a cancel. The submit button will
send an update to the server and hopefully it will all work fine.