Modules now have a year.
When adding a module, you can now choose the year you did it. When
adding grades to a module, it now only shows the ones that come from
the year you did. This means that different years of a module can have
different assignments in them, as that is what happens in real life.
closes #15.